+++ NEWS.md
@@ -1,8 +1,12 @@
 # mlr3tuning 0.6.0.9000
 
 - Fix `TuningInstance*$assign_result()` errors with required parameter bug.
+- Shortcuts to access `$learner()`, `$learners()`, `$learner_param_vals`,
+  `$predictions()` and `$resample_result()` from benchmark result in archive.
+- `extract_inner_tuning_results()` helper function to extract inner tuning 
+  results.
 
-# mlr3tuning 0.6.0
+ # mlr3tuning 0.6.0
 
 - `ArchiveTuning$data` is a public field now.